Added test cases, starting parse.go form scratch
