+++ structs/structs.go
@@ -1,0 +1,17 @@
+package structs
+
+
+// Struct definition for a node in the binary tree that
+// is returned by the parse function
+type Node struct {
+    Label string
+    Form string
+    Left *Node
+    Right *Node
+    Features []string
+}
+
+type State struct {
+    Node   *Node
+    Stream []string
+}
